refactor: v2 架构升级 - 统一CLI入口 + 瘦身SKILL.md + 解耦数据目录
Phase 1（安全修复）:
- configure.py: 改用 Gateway REST API 写配置，不再直接破坏 JSON5 文件
- 数据目录迁移: skill/data/ → ~/.openclaw/workspace-clawshop/1688-skill-data/
- CHANNEL_MAP 统一: 新建 scripts/_const.py，消灭 api.py/search.py 两套不一致映射

Phase 2（架构升级）:
- 新增 cli.py 统一入口，所有命令通过 cli.py <cmd> 调用
- 新增 scripts/cmd_check.py：AK状态/店铺/数据目录一键检查
- 内部模块规范命名: api.py/_auth.py → _api.py/_auth.py（旧文件保留为shim向后兼容）
- SKILL.md 从 601 行瘦身到 58 行（纯路由表）
- 业务细节迁移到 references/: search.md / publish.md / configure.md

数据存储: ~/.openclaw/workspace-clawshop/1688-skill-data/products/
  与 skill 代码完全解耦，重装/升级 skill 不丢失数据
